--- crates/utilejson/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "utilejson"
-edition.workspace = true
-version.workspace = true
-homepage.workspace = true
-documentation.workspace = true
-repository.workspace = true
-authors.workspace = true
-license.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tilejson.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-utiles-core = { path = "../utiles-core" }
