+++ crates/utiles-core/src/mbutiles/minzoom_maxzoom.rs
