+++ crates/utiles-core/src/tile_range.rs
@@ -59,7 +59,7 @@ impl TileRange {
 
     #[must_use]
     pub fn sql_where(&self, flip: Option<bool>) -> String {
-        // classic mbtiles sqlite query:
+        // classic mbutiles sqlite query:
         // 'SELECT tile_data FROM tiles WHERE zoom_level = ? AND tile_column = ? AND tile_row = ?',
         let miny = if flip.unwrap_or(true) {
             crate::fns::flipy(self.miny, self.zoom)
