+++ crates/utiles-dev/src/main.rs
@@ -66,7 +66,7 @@ fn mvt_dev() {
 
 async fn sqlite_deadpool_test() {
     println!("sqlite_deadpool_test");
-    let file = "D:\\blue-marble\\blue-marble.mbtiles.NOPE";
+    let file = "D:\\blue-marble\\blue-marble.mbutiles.NOPE";
     let mbta = utilesqlite::MbtilesAsync::open(file).await.unwrap();
 
     let tj = mbta.tilejson().await;
@@ -82,7 +82,7 @@ async fn sqlite_deadpool_test() {
 }
 
 async fn sqlxing() {
-    let file = "D:\\blue-marble\\blue-marble.mbtiles";
+    let file = "D:\\blue-marble\\blue-marble.mbutiles";
 
     let copts = SqliteConnectOptions::new()
         .filename(file)
