+++ crates/utilesqlite/src/mbtiles.rs
@@ -7,8 +7,8 @@ use tilejson::TileJSON;
 use tracing::{debug, error};
 
 use utiles_core::bbox::BBox;
-use utiles_core::mbtiles::metadata_row::MbtilesMetadataRow;
-use utiles_core::mbtiles::{metadata2tilejson, MinZoomMaxZoom};
+use utiles_core::mbutiles::metadata_row::MbtilesMetadataRow;
+use utiles_core::mbutiles::{metadata2tilejson, MinZoomMaxZoom};
 use utiles_core::tile_data_row::TileData;
 use utiles_core::{yflip, LngLat, Tile, TileLike};
